apps/virtual-assistant/requirements.txt
Flask==0.11.1
Flask-Cors==3.0.10
Flask-SocketIO==4.2.1
werkzeug<1.0
grpcio==1.25.0
grpcio-tools==1.25.0
inflect==3.0.2
numpy==1.17.4
python-engineio==3.11.1
python-socketio==4.4.0
requests==2.26.0
six==1.11.0
